requirements.txt
robotframework==7.2.2
robotframework-requests==0.9.7
pyyaml==6.0.2
requests==2.32.3
robotframework-jsonlibrary==0.5